Showing possible actions taking input for choice.
Added a function which figures out the possible moves and gives a choice
of what to do for each player.  Next step is to actually do the action
chosen.

Also changet a scanf to fgets because scanf and fgets don't cooperate
very well, and added looping to make sure all players gets assigned a
slot.

Updated main to call the newest couple of functions.